AUI-2086 DatePickerNative throws error when destroyed
When `destroy` is called, DatePickerNative throws an error when trying to detach one of its events. That's because the code that is attaching that event is wrong, using a non existing attribute `selector`, when it should be using `trigger` instead. That also causes the datepicker not to emit the `selectionChanged` event when new dates are selected, since the event is not attached properly.